requirements.txt
pillow
alive_progress
ffmpeg-progress-yield
gitpython
numpy
ruamel.yaml
python-slugify